requirements.txt
redis
RPi.GPIO
lgpio
gpiozero
paho-mqtt
python-dotenv
streamlit
streamlit-autorefresh
pandas